expression计算位置上移，确保hidden展示不会站位
